+++ Source/PlanningExtended 1.4/Source/Definitions/PlanVisibility.cs
@@ -1,0 +1,11 @@
+﻿namespace PlanningExtended
+{
+    public enum PlanVisibility
+    {
+        Invisible,
+
+        Visible,
+
+        PartiallyVisible
+    }
+}