+++ Source/PlanningExtended 1.4/Source/Patches/PatchPlaySettings.cs
@@ -2,7 +2,6 @@
 using PlanningExtended.Plans;
 using RimWorld;
 using Verse;
-using Verse.Sound;
 
 namespace PlanningExtended.Patches
 {
@@ -18,22 +17,11 @@ namespace PlanningExtended.Patches
 
             row.ToggleableIcon(ref arePlansVisible, Textures.ShowPlanToggleIcon, "PlanningExtended.Settings.ArePlansVisible.Label".Translate(), SoundDefOf.Mouseover_ButtonToggle);
 
-            CheckKeyBindingToggle(PlanningKeyBindingDefOf.Planning_TogglePlanVisibility, ref arePlansVisible);
+            if (arePlansVisible != PlanManager.ArePlansVisible)
+                PlanManager.SetIsPlanVisible(arePlansVisible);
 
-            PlanManager.SetArePlansVisible(arePlansVisible);
-        }
-
-        static void CheckKeyBindingToggle(KeyBindingDef keyBinding, ref bool value)
-        {
-            if (keyBinding.KeyDownEvent)
-            {
-                value = !value;
-
-                if (value)
-                    SoundDefOf.Checkbox_TurnedOn.PlayOneShotOnCamera(null);
-                else
-                    SoundDefOf.Checkbox_TurnedOff.PlayOneShotOnCamera(null);
-            }
+            if (PlanningKeyBindingDefOf.Planning_TogglePlanVisibility.KeyDownEvent)
+                PlanManager.ToggleIsPlanVisible(PlanDesignationType.Unknown);
         }
     }
 }