+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.3.0
+- Added quick selection of last loaded plans by right clicking 'load plan' designator
+- Added selection of opacity, texture and visibility for each plan type individually
+- Added custom keybinding for PlanDoors/Floors/Objects/Walls designators
+- Removed top right icon from PlanDoors/Floors/Objects/Walls designators
+
 ## 1.2.0
 - Added saving / loading / deleting of plans
 - Refactored shape code